+++ src/mb/IContinuousMediaProvider.h
@@ -43,8 +43,8 @@ public:
   virtual void resume (GingaSurfaceID surface) = 0;
   virtual void pause () = 0;
   virtual void stop () = 0;
-  virtual void setSoundLevel (float level) = 0;
-  virtual float getSoundLevel () = 0;
+  virtual void setSoundLevel (double level) = 0;
+  virtual double getSoundLevel () = 0;
   virtual void setProviderContent (void *content) = 0;
   virtual void *getProviderContent () = 0;
   virtual void getOriginalResolution (int *width, int *height) = 0;
